% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twk.randomly.R
\name{randomly}
\alias{randomly}
\title{randomly}
\usage{
randomly(data, method = "cv", ...)
}
\arguments{
\item{data}{Training data}

\item{method}{Name of folding strategie.
\itemize{
  \item \strong{cv} Cross-Validation. k can be choosen (Default: 5)
}}

\item{...}{further arguments depends on the method.}
}
\value{
List of indices in train data.
}
\description{
Folding strategie
}
\examples{
set.seed(123)
folds <- randomly(iris, 'cv', k=5)

}
