% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twk.randomly.R
\name{randomly}
\alias{randomly}
\title{randomly}
\usage{
randomly(data, method = "cv", ...)
}
\arguments{
\item{data}{Training data}

\item{method}{Name of folding strategie.
* cv Cross-Validation. k can be choosen (Default: 5)}

\item{...}{further arguments depends on the method}
}
\description{
Folding strategie
}
