% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{tweakr}
\alias{tweakr}
\title{tweakr}
\usage{
tweakr(train_set, params, k, func_train, func_predict, func_eval)
}
\arguments{
\item{train_set}{Training data}

\item{params}{List of parameters}

\item{k}{Number of folds.}

\item{func_train}{Function to train a model.}

\item{func_predict}{Function to predict the out of fold data.}

\item{func_eval}{Function to evaluate predictions.}
}
\description{
Parametertuning
}
\details{
Parametrtuning for custom models.
}
