% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twk.tweakr.R
\name{tweakr}
\alias{tweakr}
\title{tweakr}
\usage{
tweakr(train_set, params, func_train, k = 5, folds = NULL,
  save_path = NULL, save_freq = 10, twk_object = NULL, ...)
}
\arguments{
\item{train_set}{Training data}

\item{params}{List of parameters}

\item{func_train}{Function to train a model. The arguments must be `train` and `param` and return the fitted object.}

\item{k}{Number of folds.}

\item{folds}{custom folds.}

\item{save_path}{The path where the model are stored. (Default: NULL)}

\item{save_freq}{The frequence of model saving. (Defaut: 10)}

\item{twk_object}{tweakr object to continue training.}

\item{...}{Additional arguments for tweekr functions.}

\item{func_predict}{Function to predict the out of fold data. The arguments must be `fit` and `test` and return the predicted values.}

\item{func_eval}{Function to evaluate predictions. The arguments must be `pred` and `test` and return a single metric.}
}
\description{
Parametertuning
}
\details{
Parametrtuning for custom models.
}
\examples{

library(rpart)
set.seed(123)

 twk <- tweakr(train_set = iris,
params = list(cp=c(.01,.05)),
k = 10,
func_train = function(param, train, test) {
  model <- rpart(Species~. , train, control = rpart.control(cp = param$cp))
  pred <- predict(model, test, type = "class")
  error <- sum(pred == test$Species) / nrow(test)
  list(fit=model, pred=pred, eval=error)
})

prediction <- predict(twk,
                      iris,
                      func_predict = function(fit, test) predict(fit, test, type = "prob"))

}
